+++ engine/tachyon_ui.cpp
@@ -10,7 +10,7 @@ tUIElement* Tachyon_CreateUIElement(const std::string& file) {
   return element;
 }
 
-void Tachyon_DrawUIElement(Tachyon* tachyon, const tUIElement* ui_element, uint16 x, uint16 y) {
+void Tachyon_DrawUIElement(Tachyon* tachyon, const tUIElement* ui_element, int32 x, int32 y) {
   tUIDrawCommand command = {
     .ui_element = ui_element,
     .screen_x = x,
